Unshift new inputStr
